Introduce TokClass, TokRef
`TokClass` is an abstraction over a set of one more tokens.

For instance, `Plus`, `Minus`, `LessThan` are all a part of the `TokClass` called `BinaryOperator`.

It is also perfectly valid for a `TokClass` to contain only a single token. This is useful for creating a type to represent a token, since, enum variants are not types in Rust.

`TokRef` is a typesafe handle into the tok buffer. It is parameterized on `TokClass`.

This commit also implements the null-pointer optimization for `crate::tok::tokbuf::Key`.

The commit also introduces the parse module which will be populated with the AST eventually.